requirements.txt
fastapi
uvicorn
youtube-transcript-api
openai
ebooklib
markdown
requests
yt-dlp
google-generativeai